requirements.txt
django==4.0.2
Pillow==9.0.1
djangorestframework==3.13.1
sorl-thumbnail==12.8.0